pyproject.toml
[project]
name = "phototag"
version = "0.1.0"
description = "A python photo tagging utility for adding EXIF date information to PNG files"
readme = "README.md"
authors = [
    { email = "rseward@bluestone-consulting.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.0.0",
    "click>=8.0.0",
    "tqdm>=4.0.0",
]

[project.scripts]
phototag = "phototag.cli:main_with_compat"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
